requirements.txt
openai
fpdf2
python-dotenv
streamlit
python-docx
